Script relocation 2.0
