--- src/power_grid_model_io/scripts/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: Contributors to the Power Grid Model project <powergridmodel@lfenergy.org>
-#
-# SPDX-License-Identifier: MPL-2.0
